Remove hash_id from EventVersion
EventVersion originally was expected to need hash_id for versioning to
be able to quickly determine if two versions are identical or not. This
was useful when we planned to use EventVersion for implementing the
suggestions/revision feature, where we needed to be able to check if
suggestion hash ID is equal to revision hash ID. However, event
suggestions will now become a separate, stand-alone model (with hash
ID) and EventVersion no longer needs it.