Fix Event#suggest: Do not use fill on existing suggestion
SQL Alchemy tracks any changes made to persisted models and marks them
as dirty. They do not need to be explicitly added to the session. They
will automatically be updated on the next flush. To avoid this, we need
to duplicate the existing suggestion right away and then make changes to
our copy instead.